Switched to using the Academy Award DB Id
I realized, when looking at my results for 2018, that I was missing a lot of films.  Turns out that however the records in Wikidata are created they are by no certain process, complete.  The more recent films, like Bohemian Rhapsody, only had their wins attached to the records.

Assuming that IF the movie has a Academy DB Id, it is a movie that has at the very least, been nominated